% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partier.R
\name{cityblock}
\alias{cityblock}
\title{City Block Calculation}
\usage{
cityblock(x, y)
}
\arguments{
\item{x}{n-dimensional vector representing first point, or coordinate, in n-space}

\item{y}{n-dimensional vector representing second point, or coordinate, in n-space}
}
\description{
This helper function calculates the city block (i.e. Manhattan or L1) distance between two points.
This distance is simply the sum of all distances of the individual dimensions.
}
\examples{
## Find the L1 distance between (1,2) and (4,6) in 2d space.
plot(c(1,4),c(2,6)) # visually, L1 distance should be 6
cityblock(c(1,2),c(4,6))

## Find the L1 distance between two points in 4d space
cityblock(c(1,1,1,1),c(3,3,3,3))

}
\keyword{L1,}
\keyword{block,}
\keyword{city}
\keyword{manhattan,}
\keyword{taxicab}
