% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partier.R
\name{coordinate_distances}
\alias{coordinate_distances}
\title{All possible party distances}
\usage{
coordinate_distances(policies, scale.lower = 1, scale.upper = 10)
}
\arguments{
\item{policies}{matrix representing parties' policy positions, with positions being rows and parties being columns}

\item{scale.lower}{number representing lower bound for scale that all dimensions are measured on}

\item{scale.upper}{number representing upper bound for scale that all dimensions are measured on}
}
\description{
This function generates distances between all possible n-dimensioned tuples and all inputed centers and determines "closest" centers.
For example, for two dimensions on a scale from 1 to 10, there are 100 possible 2d tuples. We could have, say, two centers,
one at (2,4) and the other at (6,8) (see example). A possible tuple could be (1,1). This function calculates and records the distance from
(1,1) to both (2,4) and (6,8) and also determines that (2,4) is closest and records that.
}
\examples{
party1<-c(2,4)
party2<-c(6,8)
partypolicies<-matrix(c(party1,party2),nrow=2)
coordinate_distances(partypolicies)
}
\keyword{distances}
