% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/democracies.R
\docType{data}
\name{democracies}
\alias{democracies}
\title{Democracy data from 2005.}
\format{
An object of class \code{data.frame} with 4408 rows and 4 columns.
}
\source{
\url{https://www.tcd.ie/Political_Science/ppmd/PPMD_summary_data_GUIDE.pdf}
}
\usage{
data(democracies)
}
\description{
Data on 100 political parties in 47 countries, rated on 38 policy dimensions on a scale
from 1 (furthest left) to 20 (furthest right).
}
\references{
Kenneth Benoit and Michael Laver. 2005. Party Policy in Modern Democracies. London: Routledge.
(\url{https://www.tcd.ie/Political_Science/ppmd/PPMD_summary_data_GUIDE.pdf})
}
\keyword{datasets}
