% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partier.R
\name{plot.PartyMatrix}
\alias{plot.PartyMatrix}
\title{Plotting Party Territories}
\usage{
\method{plot}{PartyMatrix}(
  partyData,
  col = "auto",
  dimensions = "all",
  scale.lower = 1,
  scale.upper = 10,
  ...
)
}
\arguments{
\item{partyData}{object created by coordinate_distances}

\item{col}{an array of colors, one for each party, to be used to show parties' territories. Defaults to "auto" which automatically creates color scheme}

\item{dimensions}{an array of integers specifying which dimensions a plot is desired for (corresponding to the desired rows of the matrix inputted into the coordinate_distances function). Defaults to "all"}

\item{scale.lower}{number representing lower bound for scale that all dimensions are measured on}

\item{scale.upper}{number representing upper bound for scale that all dimensions are measured on}

\item{...}{Arguments to be passed to methods. See ?plot for more information. (NEEDS TO BE FIXED)}
}
\description{
This function plots parties' "territories" based on inputted party platforms
}
\examples{
party1<-c(2,4)
party2<-c(6,8)
partypolicies<-matrix(c(party1,party2),nrow=2)
partyData<-coordinate_distances(partypolicies)
plot(partyData)
}
\keyword{cityblock}
