% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partier.R
\name{plot1d}
\alias{plot1d}
\title{Plotting Party Territories on One Dimension}
\usage{
plot1d(
  partyPos,
  colorList = "auto",
  scale.lower = 1,
  scale.upper = 10,
  dimensions = "Dimension 1",
  ...
)
}
\arguments{
\item{partyPos}{vector of party positions on one dimension}

\item{colorList}{an array of colors, one for each party, to be used to show parties' territories. Defaults to "auto" which automatically creates color scheme}

\item{scale.lower}{number representing lower bound for scale that all dimensions are measured on}

\item{scale.upper}{number representing upper bound for scale that all dimensions are measured on}

\item{dimensions}{character string describing the name of the dimension}

\item{...}{Arguments to be passed to methods. See ?plot for more information. (FIXED LATER)}
}
\description{
This helper function plots parties' "territories" on one dimension based on inputted party platforms
}
\examples{
plot1d(c(1,5,3,8,6),colorList=c("red","green","blue","purple","orange"))
}
\keyword{cityblock}
