% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partier.R
\name{plot2d}
\alias{plot2d}
\title{Plotting Party Territories on Two Dimensions}
\usage{
plot2d(
  partyPos,
  colorList = "auto",
  scale.lower = 1,
  scale.upper = 10,
  dimensions = c("Dimension 1", "Dimension 2")
)
}
\arguments{
\item{partyPos}{matrix of party positions, with dimensions represented by rows and parties represented by columns}

\item{colorList}{an array of colors, one for each party, to be used to show parties' territories. Defaults to "auto" which automatically creates color scheme}

\item{scale.lower}{number representing lower bound for scale that all dimensions are measured on}

\item{scale.upper}{number representing upper bound for scale that all dimensions are measured on}

\item{dimensions}{character string array describing the names of the dimensions}

\item{...}{Arguments to be passed to methods. See ?plot for more information.}
}
\description{
This function creates a 2d plot of party territories, with one dimension represented on the x axis and the other on the y axis
}
\examples{
party1<-c(2,4)
party2<-c(6,8)
partypolicies<-matrix(c(party1,party2),nrow=2)
plot2d(partypolicies)
}
\keyword{cityblock}
