% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partier.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Plotting Territory Matrix}
\usage{
plot_matrix(partyPos, col = "auto", scale.lower = 1, scale.upper = 10, ...)
}
\arguments{
\item{partyPos}{matrix of party positions, with dimensions represented by rows and parties represented by columns}

\item{col}{an array of colors, one for each party, to be used to show parties' territories. Defaults to "auto" which automatically creates color scheme}

\item{scale.lower}{number representing lower bound for scale that all dimensions are measured on}

\item{scale.upper}{number representing upper bound for scale that all dimensions are measured on}

\item{...}{Arguments to be passed to methods. See ?plot for more information.}
}
\description{
This function creates a matrix of party territory plots based on the desired dimensions. Main diagonal plots are created using plot1d, off diagonal plots are created by plot2d
}
\examples{
party1<-c(1,2,4)
party2<-c(6,5,9)
partypolicies<-matrix(c(party1,party2),nrow=3)
plot_matrix(partypolicies)
}
\keyword{cityblock}
