% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partier.R
\name{sim_vote}
\alias{sim_vote}
\title{Generating total vote outcome summary}
\usage{
sim_vote(matrix, loss = "absolute", partypos)
}
\arguments{
\item{matrix}{output from sim_voter_positions, or another matrix with voters being rows and policy positions columns}

\item{loss}{either "absolute" or "squared" used for probability generation based on distances}

\item{partypos}{matrix of party positions with positions as rows, parties as columns}
}
\description{
This function adds a probabilistic component and generates voting outcomes for randomly generated (by sim_voter_positions) voters.
}
\examples{

}
\keyword{election,}
\keyword{generate,}
\keyword{outcomes}
\keyword{simulate,}
