% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partier.R
\name{sim_voter_positions}
\alias{sim_voter_positions}
\title{Generating voters}
\usage{
sim_voter_positions(
  num,
  range = c(0, 1),
  ndim,
  shape1,
  shape2,
  corrmat,
  rounded = FALSE
)
}
\arguments{
\item{num}{number of voters generated}

\item{range}{array of size 2 containing upper and lower bound of policy positions}

\item{ndim}{number of policy dimensions}

\item{shape1}{array corresponding to alpha parameters for beta distributions of policies}

\item{shape2}{array corresponding to beta parameters for beta distributions of policies}

\item{corrmat}{correlation matrix signifying how correlated dimensions should be. Main diagonal should be 1s}

\item{rounded}{if true, rounds voters' positions to nearest integer}
}
\description{
This function generates voters based on desired policy dimensions, scales, and underlying distributions of opinions.
}
\examples{
sim_voter_positions(10,c(1,20),3)

}
\keyword{generate,}
\keyword{simulate,}
\keyword{voters}
