% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partier.R
\name{summary.PartyMatrix}
\alias{summary.PartyMatrix}
\title{Summarizing parties' territories}
\usage{
\method{summary}{PartyMatrix}(object, ...)
}
\arguments{
\item{object}{matrix output from coordinate_distances}

\item{...}{Arguments to be passed to methods. See ?summary for more information.}
}
\description{
This function gives a well organized summary of parties' territories, including total territory counts for each party and descriptive statistics of the distances to each territory box for each party
}
\examples{
party1<-c(2,4)
party2<-c(6,8)
partypolicies<-matrix(c(party1,party2),nrow=2)
object<-coordinate_distances(partypolicies)
summary(object)
}
\keyword{summary}
